Revise transfer learning sample, after tutorial review
