--- machine-learning/tutorials/TransferLearningTF/ImageData.cs
@@ -1,16 +1,0 @@
-﻿// <SnippetAddUsings>
-using Microsoft.ML.Data;
-// </SnippetAddUsings>
-namespace TransferLearningTF
-{
-    // <SnippetDeclareTypes>
-    public class ImageData
-    {
-        [LoadColumn(0)]
-        public string ImagePath;
-
-        [LoadColumn(1)]
-        public string Label;
-    }
-    // </SnippetDeclareTypes>
-}
