--- machine-learning/tutorials/TransferLearningTF/ImagePrediction.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace TransferLearningTF
-{
-    // <SnippetDeclareTypes>
-    public class ImagePrediction : ImageData
-    {
-        public float[] Score;
-
-        public string PredictedLabelValue;
-    }
-    // </SnippetDeclareTypes>
-}
